Implement ERR_err_string using accessory ERR fns
We can pick out the individual strings of the
full error string created by ERR_err_string and
use them to write our own, identical output.

One benefit is that we don't need to create
a buffer to print. But, if we want to send the
message somewhere instead of just print it out,
then we will need one.